Refactor memory into package modules
Split the monolithic memory implementation into focused private package modules while preserving the public kitaru.memory surface and compatibility aliases for existing internal call sites. Update memory-related tests to patch the new owner modules and keep scope reset behavior pointed at the real state module.
